Add Header size functions for IPv6 and IPv4  (#569)
* Add common Header size function for IPv6 and IPv4

* MISRA: Fix rule 1.1, 10.3, 20.9 and suppress rule 20.5

+++ source/FreeRTOS_TCP_State_Handling.c
@@ -314,7 +314,7 @@
 
         if( pxTCPHeader->ucTCPFlags != 0U )
         {
-            ucIntermediateResult = uxIPHeaderSizeSocket( pxSocket ) + ipSIZE_OF_TCP_HEADER + pxTCPWindow->ucOptionLength;
+            ucIntermediateResult = ( uint8_t ) ( uxIPHeaderSizeSocket( pxSocket ) + ipSIZE_OF_TCP_HEADER + pxTCPWindow->ucOptionLength );
             xSendLength = ( BaseType_t ) ucIntermediateResult;
         }
